Added an attiny85 for programming purposes
It will be there solely for research purposes. It's clock will not be
tied in at all, it is there for the purpose of trying stuff out.

I also moved hold,hlda, and ~LOCK to the breakout header that ~AS, ALE,
and others are attached to. I have plans to move ~INT3 to the
microcontroller. I am also planning to expose the reset pin to the
i960 itself so it can do a software reset.