moved int2 and int3 to the signal breakout
I did not directly connect ~int3 to the microcontroller setup. I want to
leave it free for any future design ideas. Clock synchronization is
important but not impossible to work around.

I am still wondering if I should break the 8 most significant bits of
the address out to a separate header for future expansion effort.